Renamed folder to OpenSkyDataExtractor, refactored name usage
